Use CollectionsMarshal to avoid struct copying (#12)
* Replace long with n(u)int for calculactions

* Use CollectionsMarsha.GetValueRefOrAddDefault to avoid UnsafeSpan copying

+++ BillionChallenge/Measurements.cs
@@ -2,12 +2,12 @@
 
 public struct Measurements
 {
-    private nint _count;
+    private nuint _count;
     private nint _sum;
     private nint _min;
     private nint _max;
     
-    public double Average => _sum / _count * 0.1;
+    public double Average => _sum / (nint)_count * 0.1;
     public double Min => _min * 0.1;
     public double Max => _max * 0.1;
